requirements.txt
numpy
open3d
pynrrd
tifffile
opencv-python
scikit-image
networkx
scipy
pyvista

flask
